requirements.txt
altgraph==0.17
ansicon==1.89.0
appdirs==1.4.4
blessed==1.18.0
certifi==2020.12.5
chardet==4.0.0
flyingcircus==0.1.3.2
future==0.18.2
idna==2.10
importlib-metadata==3.7.0
jinxed==1.0.1
pefile==2019.4.18
pyinstaller==4.2
pyinstaller-hooks-contrib==2020.11
python-kucoin==2.1.2
pytk==0.0.2.1
pywin32-ctypes==0.2.0
requests==2.25.1
setuptools-scm==5.0.1
six==1.15.0
tinydb==4.4.0
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
websockets==8.1
zipp==3.4.1
